+++ app/src/main/java/eu/kanade/presentation/components/AdaptiveSheet.kt
@@ -1,9 +1,10 @@
 package eu.kanade.presentation.components
 
-import androidx.compose.animation.SizeTransform
+import androidx.activity.compose.BackHandler
 import androidx.compose.animation.core.tween
 import androidx.compose.animation.fadeIn
 import androidx.compose.animation.fadeOut
+import androidx.compose.animation.togetherWith
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.window.Dialog
@@ -27,14 +28,20 @@ fun NavigatorAdaptiveSheet(
         screen = screen,
         content = { sheetNavigator ->
             AdaptiveSheet(
-                onDismissRequest = onDismissRequest,
                 enableSwipeDismiss = enableSwipeDismiss(sheetNavigator),
+                onDismissRequest = onDismissRequest,
             ) {
                 ScreenTransition(
                     navigator = sheetNavigator,
-                    enterTransition = { fadeIn(animationSpec = tween(220, delayMillis = 90)) },
-                    exitTransition = { fadeOut(animationSpec = tween(90)) },
-                    sizeTransform = { SizeTransform() },
+                    transition = {
+                        fadeIn(animationSpec = tween(220, delayMillis = 90)) togetherWith
+                            fadeOut(animationSpec = tween(90))
+                    },
+                )
+
+                BackHandler(
+                    enabled = sheetNavigator.size > 1,
+                    onBack = sheetNavigator::pop,
                 )
             }
 
@@ -72,10 +79,10 @@ fun AdaptiveSheet(
         properties = dialogProperties,
     ) {
         AdaptiveSheetImpl(
+            modifier = modifier,
             isTabletUi = isTabletUi,
             enableSwipeDismiss = enableSwipeDismiss,
             onDismissRequest = onDismissRequest,
-            modifier = modifier,
         ) {
             content()
         }