+++ app/src/main/java/eu/kanade/presentation/more/onboarding/OnboardingScreen.kt
@@ -42,9 +42,7 @@ fun OnboardingScreen(
     }
     val isLastStep = currentStep == steps.lastIndex
 
-    BackHandler(enabled = currentStep != 0) {
-        currentStep--
-    }
+    BackHandler(enabled = currentStep != 0, onBack = { currentStep-- })
 
     InfoScreen(
         icon = Icons.Outlined.RocketLaunch,