+++ app/src/main/java/eu/kanade/tachiyomi/ui/more/OnboardingScreen.kt
@@ -33,9 +33,12 @@ class OnboardingScreen : Screen() {
 
         val restoreSettingKey = stringResource(SettingsDataScreen.restorePreferenceKeyString)
 
-        BackHandler(enabled = !shownOnboardingFlow) {
-            // Prevent exiting if onboarding hasn't been completed
-        }
+        BackHandler(
+            enabled = !shownOnboardingFlow,
+            onBack = {
+                // Prevent exiting if onboarding hasn't been completed
+            },
+        )
 
         OnboardingScreen(
             onComplete = finishOnboarding,