+++ app/src/main/java/eu/kanade/tachiyomi/ui/setting/SettingsScreen.kt
@@ -40,19 +40,19 @@ class SettingsScreen(
                     Destination.Tracking.id -> SettingsTrackingScreen
                     else -> SettingsMainScreen
                 },
-                onBackPressed = null,
-            ) {
-                val pop: () -> Unit = {
-                    if (it.canPop) {
-                        it.pop()
-                    } else {
-                        parentNavigator.pop()
+                content = {
+                    val pop: () -> Unit = {
+                        if (it.canPop) {
+                            it.pop()
+                        } else {
+                            parentNavigator.pop()
+                        }
                     }
-                }
-                CompositionLocalProvider(LocalBackPress provides pop) {
-                    DefaultNavigatorScreenTransition(navigator = it)
-                }
-            }
+                    CompositionLocalProvider(LocalBackPress provides pop) {
+                        DefaultNavigatorScreenTransition(navigator = it)
+                    }
+                },
+            )
         } else {
             Navigator(
                 screen = when (destination) {
@@ -61,7 +61,6 @@ class SettingsScreen(
                     Destination.Tracking.id -> SettingsTrackingScreen
                     else -> SettingsAppearanceScreen
                 },
-                onBackPressed = null,
             ) {
                 val insets = WindowInsets.systemBars.only(WindowInsetsSides.Horizontal)
                 TwoPanelBox(